// graphql-api/lib/interfaces/merge-report.interface.ts
export interface IPublicRecord {
  AccountDesignator: ICodeRef;
  Classification: ICodeRef;
  IndustryCode: ICodeRef;
  Status: ICodeRef;
  Type: ICodeRef;
  ExpirationDate: string | null;
  MiscPublicRecord: IMiscPublicRecord;
  FinancingStatement: IFinancingStatement;
  Garnishment: IGarnishment;
  FinancialCounseling: IFinancialCounseling;
  MaritalItem: IMaritalItem;
  Bankruptcy: IBankruptcy;
  RegisteredItem: IRegisteredItem;
  TaxLien: ITaxLien;
  LegalItem: ILegalItem;
  Foreclosure: IForeclosure;
  Remark: IRemark[];
  Source: ISource;
  subscriberCode: string | null;
  referenceNumber: string | null;
  handle: string | null;
  bureau: string | null;
  dateFiled: string | null;
  courtName: string | null;
  dateVerified: string | null;
  dateUpdated: string | null;
}
export interface IMiscPublicRecord {
  miscInformation: string | null;
}
export interface IFinancingStatement {
  CreditorType: ICodeRef;
  dateMaturity: string | null;
}
export interface IGarnishment {
  amount: number | string | null;
  dateSatisfied: string | null;
  garnishee: string | null;
  plaintiff: string | null;
}
export interface IFinancialCounseling {
  amount: number | string | null;
  dateChecked: string | null;
  dateSettled: string | null;
}
export interface IMaritalItem {
  spouse: string | null;
}
export interface IBankruptcy {
  courtNumber: string | null;
  division: string | null;
  assetAmount: number | string | null;
  dateResolved: string | null;
  exemptAmount: number | string | null;
  liabilityAmount: number | string | null;
  trustee: string | null;
  company: string | null;
  thirdParty: string | null;
}
export interface IRegisteredItem {
  Security: ICodeRef[];
  originalBalance: number | string | null;
  dateMatures: string | null;
}
export interface ITaxLien {
  amount: number | string | null;
  dateReleased: string | null;
}
export interface ILegalItem {
  CourtLocation: ICodeRef;
  CourtType: ICodeRef;
  plaintiff: string | null;
  lawyer: string | null;
  thirdParty: string | null;
  actionAmount: number | string | null;
  balance: number | string | null;
  dateSatisfied: string | null;
}
export interface IForeclosure {
  dateSettled: string | null;
  liability: number | string | null;
}